service-microsoft/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pacific-service-microsoft"
version = "0.1.0-dev"
description = "Microsoft service adapter for Pacific modules"
requires-python = ">=3.11"
license = "LicenseRef-Proprietary"
authors = [{ name = "Pacific Systems Ltd" }]
dependencies = [
    "pacific-core>=0.1.0.dev0",
    "msgraph-sdk>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pacific_service_microsoft"]

[tool.ruff]
target-version = "py311"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
